Compact LZ77 Match serialization from 9 bytes to 5 bytes
The Match struct was using u32 for both offset and length fields,
making each serialized token 9 bytes (4+4+1). Since MAX_WINDOW is
only 32768 (fits in u16) and match lengths fit in u16, this was
wasting 4 bytes per token.

Changes:
- Match fields: u32 → u16 for offset and length
- SERIALIZED_SIZE: 9 → 5 bytes (2+2+1)
- MIN_MATCH type: u32 → u16
- PreMatch internal struct: u32 → u16
- Optimal parser CostModel: updated overhead estimates for 5-byte
  tokens (literal_overhead 8→4, match_overhead 32→16 scaled bits)
- GPU dedupe: cast cl_uint → u16 at GPU/CPU boundary

This lowers the compression break-even from match length 8 to match
length 4, significantly improving compression ratios vs gzip.

+++ src/opencl.rs
@@ -2202,8 +2202,8 @@ fn dedupe_gpu_matches(gpu_matches: &[GpuMatch], input: &[u8]) -> Vec<Match> {
         };
 
         result.push(Match {
-            offset: gm.offset,
-            length: match_length as u32,
+            offset: gm.offset as u16,
+            length: match_length as u16,
             next,
         });
